Datastructure dropdown page added
