+++ src/main/resources/featureFiles/test.feature
@@ -1,15 +1,13 @@
 Feature: test
 
 @Smoke
-  Scenario: Open DsAlgo portal
+  Scenario: TC02 - Open DsAlgo portal
     Given The user opens DS Algo portal link
     Then The user should land in DS Algo portal page
+    
  @Smoke
-  Scenario: Checking Get Started link
+  Scenario: TC03- Checking Get Started link
     Given The user opens DS Algo portal link
     When The user clicks the "Get Started" button
     Then The user should be redirected to homepage
- @Smoke
-  Scenario: Checking Get Started link
-    Given The user opens DS Algo portal link
-    When The user clicks the "Get Started" button
+
