+++ src/test/java/Pages/Homepage.java
@@ -26,11 +26,6 @@ public class Homepage {
    public void clickGetStarted() {
 	getStarted.click();
 }
-   @FindBy(xpath="//a[text()='Data Structures']")
-   WebElement dataStructures;
-   
-   public void clickDataStructures() {
-	   dataStructures.click();
-   }
+  
    
 }
